requirements.txt
pymongo==4.0.1
python-telegram-bot==13.9

# Make sure to install those as additional_dependencies in the
# pre-commit hooks for pylint & mypy
certifi
# only telegram.ext: # Keep this line here; used in setup(-raw).py
tornado>=6.1
APScheduler==3.6.3
pytz>=2018.6
cachetools==4.2.2
